+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.1b1 (11.04.2022)
+
+* Bug: improved handling of minus character outside of ranges
+* Feature: only allow ASCII characters to be more compatible with the iOS implementation
+* Feature: changed error messages to be more compatible with the iOS implementation
+* Feature: changed "force" to "enforce" to be more consistent in naming
+
 ## v0.1b0 (04.04.2022)
 
 * Feature: initial implementation
+